Prevent closing of dialogs while the thread is still running.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@858 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogPromptDlg.cpp
@@ -395,6 +395,8 @@ void CLogPromptDlg::StartDiff(int fileindex)
 
 void CLogPromptDlg::OnOK()
 {
+	if (!GetDlgItem(IDOK)->IsWindowEnabled())
+		return;
 	//first add all the unversioned files the user selected
 	for (int j=0; j<m_ListCtrl.GetItemCount(); j++)
 	{
@@ -688,6 +690,8 @@ DWORD WINAPI StatusThread(LPVOID pVoid)
 
 void CLogPromptDlg::OnCancel()
 {
+	if (!GetDlgItem(IDOK)->IsWindowEnabled())
+		return;
 	DeleteFile(m_sPath);
 	UpdateData(TRUE);
 	CResizableDialog::OnCancel();
